requirements.txt
autopep8==1.4.4
entrypoints==0.3
filelock==3.0.12
fire==0.2.1
flake8==3.7.8
importlib-metadata==0.23
isort==4.3.21
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.2
opencv-python==4.1.1.26
packaging==19.2
Pillow==6.1.0
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
PyExecJS==1.5.1
pyflakes==2.1.1
pyparsing==2.4.2
six==1.12.0
termcolor==1.1.0
toml==0.10.0
tox==3.14.0
virtualenv==16.7.5
zipp==0.6.0
